refactor citation template_name / class parsing
Instead of having a function to check if a template name is a template,
normalizing it, checking against a list, and then transforming to a
"class" name to pass in to cs1.lua...

instead normalize it, and check against a dict which maps template names
to class names manually.

This makes handling corner-cases, like some of the new 'gazette' et al
templates, much easier.